Restructure internals
